requirements.txt
crewai[tools]>=0.150.0,<1.0.0
flask>=2.3.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
